refactor: Update Langfuse integration test to use getDatasetItems method with datasetName parameter
